+++ src/Symfony/Component/Console/Input/InputOption.php
@@ -156,16 +156,28 @@ class InputOption
      */
     public function acceptValue(): bool
     {
-        return $this->isValueRequired() || $this->isValueOptional();
+        return self::VALUE_NONE !== $this->valueMode();
+    }
+
+    /**
+     * Returns the option value mode.
+     */
+    public function valueMode(): int
+    {
+        return $this->mode & (self::VALUE_NONE | self::VALUE_REQUIRED | self::VALUE_OPTIONAL);
     }
 
     /**
      * Returns true if the option requires a value.
      *
      * @return bool true if value mode is self::VALUE_REQUIRED, false otherwise
+     *
+     * @deprecated since Symfony 8.2, use {@see valueMode()} instead
      */
     public function isValueRequired(): bool
     {
+        trigger_deprecation('symfony/console', '8.2', 'The "%s::isValueRequired()" method is deprecated, use "%s::valueMode()" instead.', __CLASS__, __CLASS__);
+
         return self::VALUE_REQUIRED === (self::VALUE_REQUIRED & $this->mode);
     }
 
@@ -293,8 +305,7 @@ class InputOption
             && $option->getDefault() === $this->getDefault()
             && $option->isNegatable() === $this->isNegatable()
             && $option->isArray() === $this->isArray()
-            && $option->isValueRequired() === $this->isValueRequired()
-            && $option->isValueOptional() === $this->isValueOptional()
+            && $option->valueMode() === $this->valueMode()
         ;
     }
 }